# openapi/swf-runtime.yaml
openapi: 3.0.3
info:
  title: SWF Runtime API
  version: 0.1.0
  summary: Backend-agnostic runtime protocol for SWF remote execution
  description: |
    This API mirrors the current `swf.WorkflowRuntime` boundary. It
    intentionally avoids exposing direct Strata or pgwf concepts over the wire.

    Artifact bytes are committed as part of a chapter-scoped write.
    Implementations may stage uploads internally, but artifacts are not an
    independently durable write resource.

    The normative runtime guarantees for chapter ordering, chapter
    single-assignment, lease and short-lock coordination, restart boundaries,
    and artifact visibility are documented in
    `docs/SPEC-WorkflowRuntime-Guarantees.md`.
servers:
  - url: https://swf-runtime.example.com
...
security:
  - bearerAuth: []
paths:
  /v1/tenants/{tenantId}/jobs:
    post:
      tags: [jobs]
      operationId: submitJob
      summary: Submit job
      description: |
        Submits a new job and writes the initial chapter (ordinal `0`). Any
        input artifacts are attached as part of that chapter creation. The
        server generates the job ID.
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobRequest'
      responses:
        '200':
          description: Job submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobHandle'
...
  /v1/tenants/{tenantId}/jobs/restarts:
    post:
      tags: [jobs]
      operationId: submitRestartJob
      summary: Submit restart job
      description: |
        Submits a restart derived from a prior run. Any `extraTaskInput` or
        `extraTaskOutput` artifacts are written as part of the restart chapter
        they belong to. The server generates the new job ID.
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitRestartJobRequest'
      responses:
        '200':
          description: Job restart submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobHandle'
  /v1/tenants/{tenantId}/jobs/{jobId}/restart:
    put:
      tags: [jobs]
      operationId: putRestartJob
      summary: Submit restart job with explicit job ID
      description: |
        Submits a restart using the path-supplied destination job ID. This
        operation is idempotent for equivalent requests: when the target job
        already exists with the same copied restart prefix (and restart extra
        chapter when requested), the runtime returns success; when that durable
        state exists but the internal runtime record is missing, the runtime
        repairs that record and returns success. If the existing durable state
        does not match the request, the runtime returns
        `409 existing_job_mismatch`.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitRestartJobRequest'
      responses:
        '200':
          description: Job restart submitted or confirmed/recovered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobHandle'
        '409':
          description: Existing job does not match the requested durable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tenants/{tenantId}/jobs/query:
    post:
      tags: [jobs]
      operationId: listJobs
      summary: List jobs
      description: |
        Lists jobs within the tenant identified by the path.
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListJobsRequest'
      responses:
        '200':
          description: Listed jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListJobsResponse'
  /v1/jobs/poll:
    post:
      tags: [leases]
      operationId: pollWork
      summary: Poll work
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollWorkRequest'
      responses:
        '200':
          description: Work leases
          content:
            application/json:
              schema:
                type: object
                required: [leases]
                properties:
                  leases:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionLease'
  /v1/tenants/{tenantId}/jobs/{jobId}/lease:
    post:
      tags: [leases]
      operationId: getJobLease
      summary: Get job lease
      description: |
        Attempts to acquire a lease for one specific job. The response always
        succeeds at the transport level when the lookup itself succeeds; a
        missing lease is represented as `lease: null`.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetJobLeaseRequest'
      responses:
        '200':
          description: Lease acquired or currently unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetJobLeaseResponse'
...
  /v1/tenants/{tenantId}/jobs/{jobId}/leases/{leaseId}/complete:
    post:
      tags: [leases]
      operationId: completeJobWithLease
      summary: Complete job with lease
      description: |
        Completes the current job attempt in the context of the identified
        lease.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/LeaseId'
        - $ref: '#/components/parameters/LeaseTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteExecutionRequest'
      responses:
        '204':
          description: Job completion applied
  /v1/tenants/{tenantId}/jobs/{jobId}/leases/{leaseId}/reschedule:
    post:
      tags: [leases]
      operationId: rescheduleJobWithLease
      summary: Reschedule job with lease
      description: |
        Reschedules the current job in the context of the identified lease.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/LeaseId'
        - $ref: '#/components/parameters/LeaseTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RescheduleExecutionRequest'
      responses:
        '204':
          description: Job rescheduled
  /v1/tenants/{tenantId}/jobs/{jobId}/leases/{leaseId}/add_chapter:
    post:
      tags: [leases]
      operationId: addChapterWithLease
      summary: Add chapter with lease
      description: |
        Commits a chapter in the context of the identified lease, including any
        attached artifact bytes. The target chapter ordinal is supplied in the
        request body as `chapter.ordinal`. Artifacts are chapter-scoped
        attachments, not standalone durable resources. Conflicting writes, such
        as duplicate ordinals or non-appendable ordinals, must fail rather than
        being treated as successful no-ops.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/LeaseId'
        - $ref: '#/components/parameters/LeaseTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddChapterRequest'
      responses:
        '204':
          description: Chapter stored
        '409':
          description: Conflicting chapter write
  /v1/tenants/{tenantId}/jobs/{jobId}/cancel:
    post:
      tags: [jobs]
      operationId: cancelJob
      summary: Cancel job
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelJobRequest'
      responses:
        '204':
          description: Job cancelled
  /v1/tenants/{tenantId}/jobs/{jobId}/chapters:
    get:
      tags: [chapters]
      operationId: listChapters
      summary: List chapters
      description: |
        Returns stored chapters ordered by ascending ordinal within the
        requested inclusive range. Supplying an `endOrdinal` greater than the
        highest stored ordinal is allowed and simply returns the available
        suffix. If `startOrdinal` is above the highest stored ordinal, the
        response is empty.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/StartOrdinal'
        - $ref: '#/components/parameters/EndOrdinal'
      responses:
        '200':
          description: Chapter range loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListChaptersResponse'
  /v1/tenants/{tenantId}/jobs/{jobId}/chapters/{ordinal}:
    get:
      tags: [chapters]
      operationId: getChapter
      summary: Get chapter
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/Ordinal'
      responses:
        '200':
          description: Chapter loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredChapter'
  /v1/tenants/{tenantId}/jobs/{jobId}/chapters/{ordinal}/commit-if-waiting:
    post:
      tags: [chapters]
      operationId: commitChapterIfWaiting
      summary: Commit chapter if waiting
      description: |
        Conditional chapter commit used to implement
        `TaskHandle.Finish(...)` above the runtime. The runtime acquires the
        necessary job-level coordination, verifies the supplied waiting-task
        guards, writes the chapter at `{ordinal}`, and advances the job
        atomically. If the job is no longer waiting on the described task, or
        the supplied guard fields do not match, the runtime must fail with a
        conflict response rather than silently succeeding.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/Ordinal'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitChapterIfWaitingRequest'
      responses:
        '204':
          description: Chapter commit applied
        '409':
          description: Waiting-task conflict
  /v1/tenants/{tenantId}/jobs/{jobId}/chapters/{ordinal}/artifacts/{name}:
    get:
      tags: [artifacts]
      operationId: openArtifact
      summary: Open artifact
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/Ordinal'
        - $ref: '#/components/parameters/ArtifactName'
      responses:
        '200':
          description: Artifact bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
...
    LeaseId:
      name: leaseId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    JobHandle:
      type: object
      required: [jobKey]
      properties:
        jobKey:
          $ref: '#/components/schemas/JobKey'
...
    SubmitJobRequest:
      type: object
      required: [job]
      additionalProperties: false
      properties:
        job:
          $ref: '#/components/schemas/SubmitJob'
        workerId:
          type: string
        requestTime:
          type: string
          format: date-time
...
    SubmitRestartJobRequest:
      type: object
      required: [job]
      additionalProperties: false
      properties:
        job:
          $ref: '#/components/schemas/SubmitRestartJob'
        workerId:
          type: string
        requestTime:
          type: string
          format: date-time
...